requirements.txt
langchain>=0.3.1
langchain-community>=0.3.1
langchain-core>=0.3.6
langchain-openai>=0.2.0
openai>=1.48.0
faiss-cpu>=1.8.0.post1
python-dotenv>=1.0.1
tiktoken>=0.7.0
langchain-teddynote>=0.0.34
urllib3==2.2.3
Flask==3.0.3
Flask-Cors==5.0.0